git-svn-id: https://svn.viathinksoft.com/svn/oidplus/trunk@253 02e4d621-2042-4998-a0bc-9ccd24201011

+++ ajax.php
@@ -109,7 +109,11 @@ try {
 		if ($ra_logged_in) OIDplus::authUtils()->raLogout($email);
 
 		$ra = new OIDplusRA($email);
+		if (!$ra->existing()) {
+			throw new OIDplusException("RA '$email' does not exist.");
+		}
 		$ra->delete();
+		$ra = null;
 
 		OIDplus::logger()->log("RA($email)?/A?", "RA '$email' deleted");
